Break down files; Restructure; Fix import paths.
