+++ lib/commands/HelpOptions.ts
@@ -1,4 +1,4 @@
-import { CommandPath } from '../CommandPath';
+import { CommandPath } from '../elements/CommandPath';
 
 export interface HelpOptions {
   command: undefined | CommandPath;
