+++ lib/elements/Command.ts
@@ -1,15 +1,5 @@
-import { CommandPath } from './CommandPath';
+import { CommandHandle } from './CommandHandle';
 import { OptionDefinition } from './OptionDefinition';
-import { ShowUsageFn } from './usage/ShowUsageFn';
-
-export interface CommandHandleOptions<TOptions extends {}> {
-  options: TOptions;
-  showUsage: ShowUsageFn;
-  level: number;
-  ancestors: CommandPath;
-}
-
-export type CommandHandle<TOptions extends {}> = (params: CommandHandleOptions<TOptions>) => void | Promise<void>;
 
 export interface Command<TOptions extends {}> {
   name: string;
