+++ lib/elements/CommandHandle.ts
@@ -1,0 +1,3 @@
+import { CommandHandleOptions } from './CommandHandleOptions';
+
+export type CommandHandle<TOptions extends {}> = (params: CommandHandleOptions<TOptions>) => void | Promise<void>;
