+++ lib/elements/CommandHandleOptions.ts
@@ -1,0 +1,6 @@
+export interface CommandHandleOptions<TOptions extends {}> {
+  options: TOptions;
+  showUsage: ShowUsageFn;
+  level: number;
+  ancestors: CommandPath;
+}
