+++ lib/elements/CommandPath.ts
