+++ lib/elements/OptionDefinition.ts
