+++ lib/index.ts
@@ -1,12 +1,16 @@
-import { OptionDefinition } from './OptionDefinition';
+import { Command } from './elements/Command';
+import { CommandHandle } from './elements/CommandHandle';
+import { CommandHandleOptions } from './elements/CommandHandleOptions';
+import { CommandPath } from './elements/CommandPath';
+import { OptionDefinition } from './elements/OptionDefinition';
 import { runCli } from './runCli';
 import { ShowUsageFn } from './usage/ShowUsageFn';
-import { Command, CommandHandle, CommandHandleOptions } from './Command';
 
 export {
   Command,
   CommandHandle,
   CommandHandleOptions,
+  CommandPath,
   OptionDefinition,
   runCli,
   ShowUsageFn
