+++ lib/recommend/RecommendCommandFn.ts
@@ -1,4 +1,4 @@
-import { CommandPath } from '../CommandPath';
+import { CommandPath } from '../elements/CommandPath';
 
 export type RecommendCommandFn = (params: {
   commandPath: CommandPath;
