+++ lib/recommend/getRecommendCommand.ts
@@ -1,4 +1,4 @@
-import { Command } from '../Command';
+import { Command } from '../elements/Command';
 import { recommendCommand } from './recommendCommand';
 import { RecommendCommandFn } from './RecommendCommandFn';
