+++ lib/runCli.ts
@@ -1,5 +1,5 @@
 import { addHelpCommandToCli } from './addHelpCommandToCli';
-import { Command } from './Command';
+import { Command } from './elements/Command';
 import { getRecommendCommand } from './recommend/getRecommendCommand';
 import { getShowUsage } from './usage/getShowUsage';
 import { runCliRecursive } from './runCliRecursive';
