+++ lib/usage/ShowUsageFn.ts
@@ -1,4 +1,4 @@
-import { CommandPath } from '../CommandPath';
+import { CommandPath } from '../elements/CommandPath';
 
 export type ShowUsageFn = (params: {
   commandPath: CommandPath;
